requirement.txt
ConfigArgParse==1.7
einops==0.8.0
imageio==2.35.1
ipdb
lpips==0.1.4
matplotlib==3.7.5
matplotlib-inline==0.1.7
numpy==1.24.4
open3d==0.18.0
opencv-python==4.9.0.80
pandas==2.0.3
pillow==10.2.0
plotly==5.21.0
protobuf==5.26.1
PyYAML==6.0.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
spconv-cu118==2.3.6
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tqdm==4.66.2
trimesh==4.3.1
yacs==0.1.8
